+++ src/js/validator/priceValiator.js
